Moved and renamed `isStoredInSecretStorage`
