+++ src/secret-storage.ts
@@ -672,30 +672,3 @@ export function trimTrailingEquals(input: string): string {
         return input;
     }
 }
-
-/**
- * Check whether the private keys exist in secret storage.
- *
- * @param secretStorage - The secret store using account data
- * @returns map of key name to key info the secret is encrypted
- *     with, or null if it is not present or not encrypted with a trusted
- *     key
- */
-export async function isStoredInSecretStorage(
-    secretStorage: ServerSideSecretStorage,
-): Promise<Record<string, object> | null> {
-    // check what SSSS keys have encrypted the master key (if any)
-    const stored = (await secretStorage.isStored("m.cross_signing.master")) || {};
-    // then check which of those SSSS keys have also encrypted the SSK and USK
-    function intersect(s: Record<string, SecretStorageKeyDescription>): void {
-        for (const k of Object.keys(stored)) {
-            if (!s[k]) {
-                delete stored[k];
-            }
-        }
-    }
-    for (const type of ["self_signing", "user_signing"]) {
-        intersect((await secretStorage.isStored(`m.cross_signing.${type}`)) || {});
-    }
-    return Object.keys(stored).length ? stored : null;
-}
